requirements.txt
flask
cloudpickle
googleappenginecloudstorageclient
google-cloud-storage